Delete with dependencies updated to add a check if the data is empty skip

+++ app/Http/Services/BookTypeService.php
@@ -107,10 +107,13 @@ class BookTypeService
 			// Finding related books
 			$books = $this->booksService->getAllWithBookType($id);
 
-			// Delete related books
-			foreach ($books['data'] as $book)
+			if (!empty($books['data']))
 			{
-				$this->booksService->delete($book->id);
+				// Delete related books
+				foreach ($books['data'] as $book)
+				{
+					$this->booksService->delete($book->id);
+				}
 			}
 
 			// Delete book type
